requirements.txt
aiohttp>=3.9,<4
python-socketio[client]>=5.11,<6
